services/agent_swarm/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic>=0.40.0
psycopg2-binary==2.9.9
requests==2.32.3
pydantic==2.10.6
python-dotenv==1.0.1
fal-client>=0.10.0,<1.0.0
google-cloud-storage>=2.0.0
openai>=1.50.0
